requirements.txt
pyyaml
numpy
scipy
pymatgen
tqdm